move channel attributes to a submodule
